refactor: change travel modal to ride modal
