+++ components/modal/travelModal/EditRideDirectionModal.tsx
@@ -3,22 +3,21 @@ import { TextInputBase } from "@/components/input/TextInput"
 import ModalTemplate from "@/components/ModalTemplate"
 import { useTheme } from "@/context/ThemeContext"
 import { modalElementStyles, modalStyles } from "@/src/styles/ModalStyles"
-import { EditableTravelRouteModalProp } from "@/src/types/EditableTravels"
-import { Route } from "@/src/types/Travels"
+import { EditableTravelDirectionModalProp } from "@/src/types/EditableTravels"
 import { useMemo } from "react"
 import { Pressable, View } from "react-native"
 import FlatlistBase from "../FlatlistPicker"
 
-export default function EditTravelRouteModal({ routes, searchQuery, isModalVisible, setSearchQuery, onClose, onSelect }: EditableTravelRouteModalProp) {
+export default function EditRideDirectionModal({ directions, searchQuery, isModalVisible, setSearchQuery, onClose, onSelect }: EditableTravelDirectionModalProp) {
     const { theme } = useTheme()
 
     const filteredItems = useMemo(() => {
-        if (!routes) return []
+        if (!directions) return []
         const query = searchQuery.toLowerCase()
-        return routes.filter(route =>
-            route.name.toLowerCase().includes(query) || route.code.toLowerCase().includes(query) || route.vehicle_type.name.toLowerCase().includes(query)
+        return directions.filter(direction =>
+            direction.name.toLowerCase().includes(query)
         )
-    }, [routes, searchQuery])
+    }, [directions, searchQuery])
 
     return (
         <ModalTemplate.Bottom
@@ -27,14 +26,14 @@ export default function EditTravelRouteModal({ routes, searchQuery, isModalVisib
         >
             <ModalTemplate.BottomContainer>
                 <View style={modalElementStyles[theme].header}>
-                    <Input.Header>Select a route</Input.Header>
+                    <Input.Header>Select a direction</Input.Header>
                     <Pressable onPress={onClose}>
                         <Input.Subtitle>Close</Input.Subtitle>
                     </Pressable>
                 </View>
                 <TextInputBase.Clear
                     value={searchQuery}
-                    placeholder="Search route..."
+                    placeholder="Search direction..."
                     onChangeText={setSearchQuery}
                     onClear={() => setSearchQuery('')}
                 />
@@ -47,11 +46,8 @@ export default function EditTravelRouteModal({ routes, searchQuery, isModalVisib
                         items={filteredItems}
                         onSelect={onSelect}
                     >
-                        {(item: Route) => (
-                            <FlatlistBase.PickerItem item={item}>
-                                <Input.ValueText>{item.code}</Input.ValueText>
-                                <Input.ValuePrimary>{item.name}</Input.ValuePrimary>
-                            </FlatlistBase.PickerItem>
+                        {(item) => (
+                            <Input.Label>{item.name}</Input.Label>
                         )}
                     </FlatlistBase.Picker>
                 )}
